Remove connector properly

+++ httpdocs/model/connector.delete.php
@@ -18,7 +18,6 @@ $uid = $ses->isSessionValid() ? $ses->getId() : 0;
 
 $cid = isset( $_REQUEST[ 'cid' ] ) ? intval( $_REQUEST[ 'cid' ] ) : 0;
 $ci_type = isset( $_REQUEST[ 'class_instance_type' ] ) ? $_REQUEST[ 'class_instance_type' ] : 'none';
-$ci_relation = isset( $_REQUEST[ 'class_instance_relation' ] ) ? $_REQUEST[ 'class_instance_relation' ] : 'none';
 
 if ( $pid )
 {
@@ -28,9 +27,13 @@ if ( $pid )
     if(!$cit_id) { echo makeJSON( array( '"error"' => 'Can not find "'.$ci_type.'" class for this project' ) ); return; }
     
     // relation ids
-    $cir_id = $db->getRelationId( $pid, $ci_relation );
-    if(!$cir_id) { echo makeJSON( array( '"error"' => 'Can not find "'.$ci_relation.'" relation for this project' ) ); return; }
+    $cir_id = $db->getRelationId( $pid, 'model_of' );
+    if(!$cir_id) { echo makeJSON( array( '"error"' => 'Can not find "model_of" relation for this project' ) ); return; }
 
+    // for labels, only remove the relation
+    $lab_id = $db->getRelationId( $pid, 'labeled_as' );
+    if(!$lab_id) { echo makeJSON( array( '"error"' => 'Can not find "labeled_as" relation for this project' ) ); return; }
+       
     // retrieve class instance id
     $classin = $db->getResult('SELECT "cci"."class_instance_id" AS "id" FROM "connector_class_instance" AS "cci"
      WHERE "cci"."relation_id" = '.$cir_id.' AND "cci"."connector_id" = '.$cid.' AND 
@@ -49,11 +52,37 @@ if ( $pid )
     {
       
       // XXX: correct deletion of associated terminals
-      
+      $presyn_id = $db->getRelationId( $pid, "presynaptic_to" );
+      if(!$presyn_id) { echo makeJSON( array( '"error"' => 'Can not find "presynaptic_to" relation for this project' ) ); return; }
+      $postsyn_id = $db->getRelationId( $pid, "postsynaptic_to" );
+      if(!$postsyn_id) { echo makeJSON( array( '"error"' => 'Can not find "postsynaptic_to" relation for this project' ) ); return; }
+
+      // retrieve and delete pre and post terminal
+      $conin = $db->getResult('SELECT "cici"."class_instance_a" AS "id" FROM "class_instance_class_instance" AS "cici"
+       WHERE ("cici"."relation_id" = '.$presyn_id.' OR "cici"."relation_id" = '.$postsyn_id.')
+       AND "cici"."class_instance_b" = '.$classin_id.' AND 
+       "cici"."project_id" = '.$pid);
+       
       // delete connector
       $ids = $db->deleteFrom("connector", ' "connector"."id" = '.$cid);
-      // delete class_instance
+      
+      // delete class_instance      
       $ids = $db->deleteFrom("class_instance", ' "class_instance"."id" = '.$classin_id);
+      
+      // delete class_instance
+      if(!empty($conin)) {
+        foreach($conin as $key => $tn) {
+          echo "delete ci".$tn['id'];
+           $ids = $db->deleteFrom("class_instance", ' "class_instance"."id" = '.$tn['id']);
+        }
+      }
+      
+      // delete label relationships without deleting the class_instance labels
+      $ids = $db->deleteFrom("treenode_class_instance", ' "treenode_class_instance"."treenode_id" = '.$tnid.' AND
+      "treenode_class_instance"."relation_id" = '.$lab_id);
+
+      echo "Removed connector successfully.";
+          
     } else {
       echo makeJSON( array( '"error"' => 'Can not delete. You are not the owner of the class_instance "'.$classin_id.'" for this project' ) );
       return;
